chore: simplify browser test
